Added a main test class, updated the gradle build
Reconfigured the tests to all run consecutively, and import assets correctly.

Edited the Gradle config so that AllTests is run and then a report is generated.

Created a runnable jar file which can be executed from the command line.

Tidied some test imports.
