+++ .gitignore
@@ -125,3 +125,6 @@ Thumbs.db
 !/ios-moe/xcode/*.xcodeproj/xcshareddata
 !/ios-moe/xcode/*.xcodeproj/project.pbxproj
 /ios-moe/xcode/native/
+!/tests/build/
+/tests/build/*
+!tests/build/reports
