+++ build.gradle
@@ -69,31 +69,13 @@ project(":core") {
 
 
 project(":tests") {
-    apply plugin: "java-library"
-
-    sourceSets.test.java.srcDirs = ["src/"]
-
-    dependencies {
-
-        compile project(":core")
-
-        compile "junit:junit:4.+"
-        compile "org.mockito:mockito-all:1.9.+"
-
-        compile "com.badlogicgames.gdx:gdx-backend-headless:$gdxVersion"
-        compile "com.badlogicgames.gdx:gdx:$gdxVersion"
-
-        testCompile 'junit:junit:4.+'
-        testCompile "org.mockito:mockito-all:1.9.+"
-
-        testCompile "com.badlogicgames.gdx:gdx-backend-headless:$gdxVersion"
-        testCompile "com.badlogicgames.gdx:gdx:$gdxVersion"
-        testCompile "com.badlogicgames.gdx:gdx-platform:$gdxVersion:natives-desktop"
-    }
-}
-
-project(":tests") {
     apply plugin: "java"
+    
+    test {
+    	filter {
+    		includeTestsMatching "com.dicycat.kroy.AllTests"
+    	}
+    }
 
     sourceSets.test.java.srcDirs = ["src/"]
