+++ tests/build.gradle
@@ -3,6 +3,6 @@ apply plugin: "java"
 sourceCompatibility = 1.8
 [compileJava, compileTestJava]*.options*.encoding = 'UTF-8'
 
-sourceSets.main.java.srcDirs = [ "src/" ]
+sourceSets.main.java.srcDirs = [ "../core/src/" ]
 
 project.ext.assetsDir = [ "../core/assets/" ]
